reorganize members code
